scripts/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
google-genai>=1.0.0
boto3>=1.28.0
apify-client>=1.0.0
Pillow>=10.0.0
beautifulsoup4>=4.13.0
playwright>=1.54.0
playwright-stealth>=1.0.6
